refactor duplications
