+++ apps/modernization-api/src/main/java/gov/cdc/nbs/patient/profile/investigation/AddInvestigationRedirector.java
@@ -1,9 +1,6 @@
 package gov.cdc.nbs.patient.profile.investigation;
 
-import gov.cdc.nbs.patient.profile.redirect.ReturningPatientCookie;
-import gov.cdc.nbs.patient.profile.redirect.outgoing.ClassicPatientProfilePreparer;
-import gov.cdc.nbs.patient.profile.redirect.outgoing.ClassicPatientSearchPreparer;
-import org.springframework.http.HttpStatus;
+import gov.cdc.nbs.patient.profile.redirect.outgoing.ClassicPatientProfileRedirector;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -20,36 +17,23 @@ class AddInvestigationRedirector {
 
     private static final String LOCATION = "/nbs/LoadSelectCondition1.do";
 
-    private final ClassicPatientSearchPreparer searchPreparer;
-    private final ClassicPatientProfilePreparer profilePreparer;
+    private final ClassicPatientProfileRedirector redirector;
 
-    AddInvestigationRedirector(
-        final ClassicPatientSearchPreparer searchPreparer,
-        final ClassicPatientProfilePreparer profilePreparer
-    ) {
-        this.searchPreparer = searchPreparer;
-        this.profilePreparer = profilePreparer;
+    AddInvestigationRedirector(final ClassicPatientProfileRedirector redirector) {
+        this.redirector = redirector;
     }
 
     @PreAuthorize("hasAuthority('ADD-INVESTIGATION')")
     @GetMapping("/nbs/api/profile/{patient}/investigation")
     ResponseEntity<Void> add(@PathVariable("patient") final long patient) {
-        prepare(patient);
 
         URI location = UriComponentsBuilder.fromPath(LOCATION)
             .queryParam("ContextAction", "AddInvestigation")
             .build()
             .toUri();
 
-        return ResponseEntity
-            .status(HttpStatus.CREATED)
-            .location(location)
-            .headers(new ReturningPatientCookie(patient)::apply)
-            .build();
+        return redirector.preparedRedirect(patient, location);
     }
 
-    private void prepare(final long patient) {
-        searchPreparer.prepare();
-        profilePreparer.prepare(patient);
-    }
+
 }
